+++ globalgiving/commands/cmd_crawled.py
@@ -2,13 +2,19 @@ import click
 import requests
 from googlesearch import search
 from globalgiving.cli import pass_context
-from globalgiving.crawler.crawl_functions import rank_all, url_rank
 from urllib.parse import urlparse
 import dotenv
 import os
+import sys
 import pymongo
 from operator import itemgetter
 import json
+
+SCRAPER_REG_PATH = "../../../microservices"  # Sibling package path
+
+# Bring microservices directory into import path
+sys.path.append(os.path.realpath(os.path.dirname(__file__) + SCRAPER_REG_PATH))
+from scraper_crawler.crawl_functions import rank_all, url_rank
 
 
 @click.command("crawled", short_help="Crawl for new directories and NGOs")
