+++ globalgiving/requirements.txt
@@ -5,6 +5,4 @@ tabulate==0.8.2
 pymongo==3.7.2
 requests==2.20.0
 PyJWT==1.6.4
-google==2.0.1
-phonenumbers
 cookiecutter==1.6.0
