+++ microservices/scraper_crawler/requirements.txt
@@ -1,0 +1,4 @@
+beautifulsoup4==4.6.3
+google==2.0.1
+phonenumbers
+requests==2.20.0
